requirements.txt
cffi==1.1.2
click==4.0
coloredlogs==1.0.1
cryptography==0.9.1
enum34==1.0.4
humanfriendly==1.32
idna==2.0
ipaddress==1.0.7
ndg-httpsclient==0.4.0
nose==1.3.7
pyasn1==0.1.8
pycparser==2.14
pyOpenSSL==0.15.1
python-dotenv==0.1.2
requests==2.7.0
six==1.9.0
